embodied-red-teaming/a_evalUnderAttack_noTraining/baseline_run.py
#!/usr/bin/env python3
# main.py
"""
ERT main loop:
- 加载任务->image links
- 调用 generate_instructions.EmbodiedRedTeamModel 生成 candidate instructions
- 在 OpenVLA + LIBERO 上评估每条 instruction 的 success_rate
- 选失败示例反馈给生成器，迭代若干轮
- 输出最终结果到 YAML/JSON
"""
import os
os.environ['CUDA_VISIBLE_DEVICES'] = '7'
import warnings
warnings.filterwarnings("ignore", category=RuntimeWarning)
import os
import argparse
import yaml
import json
import random
import time
import hashlib
from functools import partial
from multiprocessing import Pool
from tqdm import tqdm
from dataclasses import dataclass
import draccus
from typing import Optional, Union
from pathlib import Path
import re
from datetime import datetime

import sys
sys.path.insert(0, "/home/baoshuntong/code/saftyEmbodyAI/embodied-red-teaming/LIBERO")
sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))

# 获取当前时间
now = datetime.now()

# 格式化为字符串，例如：2025-10-05_15-36-22
NOW_TIME_STR = now.strftime("%Y-%m-%d_%H-%M-%S")
from libero.libero import benchmark
from a_evalUnderAttack_noTraining.utils import eval_libero
from embodyRedTeaming_baseline.generate_instructions import EmbodiedRedTeamModelWithQwen2, CLIPEmbeddingModel
from openvla.experiments.robot.openvla_utils import get_processor

from openvla.experiments.robot.robot_utils import (
    get_model,
)
from openvla.experiments.robot.libero.libero_utils import (
    get_libero_env,
)

@dataclass
class GenerateConfig:
    # fmt: off

    #################################################################################################################
    # Model-specific parameters
    #################################################################################################################
    model_family: str = "openvla"                    # Model family
    pretrained_checkpoint: Union[str, Path] = "openvla/openvla-7b-finetuned-libero-spatial"     # Pretrained checkpoint path
    load_in_8bit: bool = False                       # (For OpenVLA only) Load with 8-bit quantization
    load_in_4bit: bool = False                       # (For OpenVLA only) Load with 4-bit quantization

    center_crop: bool = True                         # Center crop? (if trained w/ random crop image aug)

    #################################################################################################################
    # LIBERO environment-specific parameters
    #################################################################################################################
    task_suite_name: str = "libero_spatial"          # Task suite. Options: libero_spatial, libero_object, libero_goal, libero_10, libero_90
    num_steps_wait: int = 10                         # Number of steps to wait for objects to stabilize in sim
    num_trials_per_task: int = 10                    # Number of rollouts per task

    #################################################################################################################
    # Utils
    #################################################################################################################
    run_id_note: Optional[str] = None                # Extra note to add in run ID for logging
    local_log_dir: str = "./experiments/logs"        # Local directory for eval logs

    use_wandb: bool = False                          # Whether to also log results in Weights & Biases
    wandb_project: str = "redTeamingOpenvla"        # Name of W&B project to log to (use default!)
    wandb_entity: str = "tongbs-sysu"          # Name of entity to log under

    seed: int = 7                                    # Random Seed (for reproducibility)

    # user-customized config

    is_save_video: bool = False                       # Whether to save rollout video

    task_to_huglinks_json_path: str = "libero-init-frames/json_data_for_rl/vlm_initial_state_links.json"  # JSON file mapping task names to image URLs

    redTeaming_vlm_model = "qwen2.5-vl-72b-instruct"               # VLM model for instruction generation (e.g. gpt-4, gpt-3.5-turbo)

    num_rejection_samples: int = 5               # Number of rejection samples M for EmbodiedRedTeamModel

    num_instructions : int = 20                     # Number of instructions to request per generation

    failure_threshold: float = 0.5               # Success rate <= threshold considered a failure

    max_feedback_examples: int = 10                  # Max failure examples to feed back each round

    output_path: str = f"./output/{NOW_TIME_STR}/baseline_redteaming_results.json"  # Path to save the attack results (YAML)

    examples_path: str = ""  # Path to existing examples (YAML), optional

    n_iter_attack: int = 1                           # Number of attack iterations per task


def parse_task_and_links(cfg,task_to_links):

    task_suite_name = cfg.task_suite_name
    
    task_suite = task_to_links[task_suite_name]

    task_language_list = task_suite.keys()

    return task_language_list, task_suite


# -------------- 主循环 --------------
@draccus.wrap()
def run_ert_loop(cfg: GenerateConfig):
    # load task images
    with open(cfg.task_to_huglinks_json_path, "r") as f:
        task_to_links = json.load(f)  # {task: [image_url, ...]}

    # load examples
    examples = {}
    if cfg.examples_path and os.path.exists(cfg.examples_path):
        with open(cfg.examples_path, "r") as f:
            examples = yaml.safe_load(f) 
    
            # [OpenVLA] Get Hugging Face processor
    processor = None
    if cfg.model_family == "openvla":
        processor = get_processor(cfg)
    openvla_model = get_model(cfg)

    # prepare embedding & red team generator
    embedding_model = CLIPEmbeddingModel(device="cuda")
    red_team = EmbodiedRedTeamModelWithQwen2(embedding_model=embedding_model, model=cfg.redTeaming_vlm_model,
                                    num_rejection_samples=cfg.num_rejection_samples)

    task_language_list, task_links_suite = parse_task_and_links(cfg,task_to_links)

    benchmark_dict = benchmark.get_benchmark_dict()
    task_suite = benchmark_dict[cfg.task_suite_name]()
    # num_tasks_in_suite = task_suite.n_tasks
    print(f"Task suite: {cfg.task_suite_name}")

    suite_task_num = 0

    suite_task_success = 0


    results = []  # task -> list of dicts per round / per candidate

    for task_id,task_language in tqdm(enumerate(task_language_list), desc="Tasks"):

        task_examples = examples.get(task_language, [])
        
        if task_examples:
            in_context_examples = []
            for task_example in task_examples:
                    current_example = task_example['failed_examples']
                    in_context_examples.extend(current_example)
        else:
            in_context_examples = task_examples

        instruction = task_language.replace("_", " ")


        task = task_suite.get_task(task_id)
        # Get default LIBERO initial states
        initial_states = task_suite.get_task_init_states(task_id)
        # Initialize LIBERO environment and task description
        env, _ = get_libero_env(task, cfg.model_family, resolution=256)
        

        # pick an image_url (initial frame) if use_image

        current_task_links = task_links_suite[task_language]

        image_url = random.choice(current_task_links)

        # generate instruction group (EmbodiedRedTeamModel handles rejection sampling)
        for _ in range(cfg.n_iter_attack):  # 每个任务最多尝试 n 轮

            annotations = red_team(instruction, image_url=image_url, examples=in_context_examples,
                                num_instructions=cfg.num_instructions,select_topk=5)
            
            task_results = {}
            done_results = []
            failed_examples = []
            for i,annotation in enumerate(annotations):
                match = re.search(r'^\s*\d+\.\s*"(.*?)"', annotation)
                if match:
                    annotation = match.group(1)
                    print(instruction)
                print(f"Use instruction: {annotation} to attack task {instruction}")
                task_episodes, task_successes = eval_libero(cfg, openvla_model,processor,env,initial_states,annotation)
                # eval_results.append(eval_result)

                suite_task_success += task_successes
                suite_task_num += task_episodes

                success_rate = task_successes / task_episodes

                if success_rate <= cfg.failure_threshold:
                    failed_examples.append(annotation)
                else:
                    done_results.append(annotation)
                
            if len(failed_examples) > cfg.max_feedback_examples:
                failed_examples = failed_examples[-cfg.max_feedback_examples:]
                in_context_examples = list(set(in_context_examples + failed_examples))  # unique``
        
        task_results['done_annotations'] = done_results 
        task_results['fail_annotations'] = failed_examples
        task_results['image_url'] = image_url
        task_results['task'] = instruction

        results.append(task_results)

        # save intermediate results to disk after each task
        os.makedirs(os.path.dirname(cfg.output_path) or ".", exist_ok=True)
        with open(cfg.output_path, "w") as fout:
            json.dump(results, fout,indent=4)
    
    print(f"Overall success rate on {suite_task_num} episodes: {suite_task_success / suite_task_num:.3f}")

    print(f"Done. Results saved to {cfg.output_path}")


if __name__ == "__main__":

    run_ert_loop()
