embodied-red-teaming/requirements_vla.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiohttp-cors==0.8.1
aiosignal==1.4.0
airportsdata==20250909
annotated-types==0.7.0
anthropic==0.69.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
array_record==0.8.1
astor==0.8.1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
av==16.0.1
bddl==3.6.0
blake3==1.0.8
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.2.1
cloudpickle==3.1.1
codetiming==1.4.0
colorful==0.5.7
compressed-tensors==0.9.3
contourpy==1.3.2
cuda-bindings==13.0.2
cuda-pathfinder==1.3.1
cuda-python==13.0.2
cupy-cuda12x==13.6.0
cycler==0.12.1
datasets==4.2.0
decorator==5.2.1
decord==0.6.0
Deprecated==1.2.18
depyf==0.18.0
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dlimp @ git+https://github.com/moojink/dlimp_openvla@040105d256bd28866cc6620621a3d5f7b6b91b46
dm-tree==0.1.9
dnspython==2.8.0
docstring_parser==0.17.0
draccus==0.8.0
easydict==1.13
einops==0.8.1
email-validator==2.3.0
etils==1.13.0
evdev==1.9.2
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.119.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
fastjsonschema==2.21.2
fastrlock==0.8.3
fastuuid==0.13.5
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
future==1.0.0
gast==0.6.0
gguf==0.17.1
gitdb==4.0.12
GitPython==3.1.45
glfw==2.10.0
google-api-core==2.26.0
google-auth==2.41.1
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.75.1
gym==0.26.2
gym-notices==0.1.0
h11==0.16.0
h5py==3.15.0
hf-xet==1.1.10
hf_transfer==0.1.9
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.35.3
hydra-core==1.3.2
identify==2.6.15
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.0.0
importlib_resources==6.5.2
iniconfig==2.1.0
interegular==0.3.3
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
json-numpy==2.1.1
jsonlines==4.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.8.1
jupytext==1.17.3
keras==2.15.0
kiwisolver==1.4.9
lark==1.2.2
libclang==18.1.1
liger_kernel==0.6.2
litellm==1.78.0
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.12
loguru==0.7.3
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mathruler==0.1.0
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdit-py-plugins==0.5.0
mdurl==0.1.2
megatron-core @ git+https://github.com/NVIDIA/Megatron-LM.git@64cb6bd7a155c4e742514ed1024407fb97d0a367
mergedeep==1.3.4
mistral_common==1.8.5
ml-dtypes==0.2.0
modelscope==1.31.0
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.19.0
mujoco==3.3.7
multidict==6.7.0
multiprocess==0.70.16
mypy_extensions==1.1.0
nanobind==2.9.2
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.13.0
nltk==3.9.2
nodeenv==1.9.1
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
omegaconf==2.3.0
openai==2.3.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-fixer==0.2.5
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
OpenEXR==3.4.1
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-exporter-prometheus==0.58b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-semantic-conventions-ai==0.4.13
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.3
outlines==0.1.11
outlines_core==0.1.26
packaging==25.0
pandas==2.3.3
parso==0.8.5
partial-json-parser==0.2.1.1.post6
peft==0.11.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
flash-attn==2.5.5
pluggy==1.6.0
pre_commit==4.3.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
promise==2.3
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.21.12
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==3.0.1
pycountry==24.6.1
pycparser==2.23
pydantic==2.12.2
pydantic-extra-types==2.10.6
pydantic_core==2.41.4
pyext==0.7
Pygments==2.19.2
pylatexenc==2.10
pynput==1.8.1
pynvml==13.0.1
PyOpenGL==3.1.10
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0
python-multipart==0.0.20
python-xlib==0.33
pytz==2025.2
PyYAML==6.0.3
pyyaml-include==1.4.1
pyzmq==27.1.0
qwen-vl-utils==0.0.14
ray==2.50.0
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.0
robosuite==1.4.1
rpds-py==0.27.1
rsa==4.9.1
ruff==0.14.0
safetensors==0.6.2
scipy==1.15.3
sentencepiece==0.1.99
sentry-sdk==2.42.0
setproctitle==1.3.7
sgl-kernel==0.1.0
sglang==0.4.6.post1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
stack-data==0.6.3
starlette==0.48.0
sympy==1.13.1
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensordict==0.6.2
tensorflow==2.15.0
tensorflow-addons==0.23.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.15.0
tensorflow-graphics==2021.12.3
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.17.2
termcolor==3.1.0
tiktoken==0.12.0
timm==0.9.10
tokenizers==0.22.1
toml==0.10.2
tomli==2.3.0
torch==2.2.0
torch_memory_saver==0.0.8
torchaudio==2.2.0
torchdata==0.11.0
torchvision==0.17.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
trimesh==4.8.3
# triton==2.2.0
typeguard==2.13.3
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
virtualenv==20.35.3
vllm==0.8.5.post1
wandb==0.22.2
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.14.2
xformers==0.0.29.post2
xgrammar==0.1.18
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
